SOURCE_CODE_EXTRACTED/integration/shared-auth/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pyjwt[crypto]==2.9.0
cryptography==44.0.0
pydantic==2.10.4
python-multipart==0.0.20
